refactor: 꼬리질문 생성 컨트롤러 interview로 옮김
